+++ shared/src/commonMain/kotlin/com/ekdorn/classicdungeon/shared/maths/Vector.kt
@@ -3,17 +3,17 @@ package com.ekdorn.classicdungeon.shared.maths
 import kotlin.math.sqrt
 
 
-internal data class Vector (var x: Double, var y: Double) {
-    constructor (): this(0.0, 0.0)
-    constructor (x: Int, y: Int): this(x.toDouble(), y.toDouble())
+internal data class Vector (var x: Float, var y: Float) {
+    constructor (): this(0F, 0F)
+    constructor (x: Int, y: Int): this(x.toFloat(), y.toFloat())
 
     inline var w: Int
         get () = x.toInt()
-        set (v) { x = v.toDouble() }
+        set (v) { x = v.toFloat() }
 
     inline var h: Int
         get () = y.toInt()
-        set (v) { y = v.toDouble() }
+        set (v) { y = v.toFloat() }
 
     fun rangeTo (other: Vector) = Vector(x - other.x, y - other.y)
     fun length () = sqrt(x * x + y * y)
